Extract shared test support utilities
