+++ tests/support/__init__.py
@@ -1,0 +1,6 @@
+"""Shared support utilities for tests.
+
+Modules under this package are test infrastructure, not product code. Keep
+them focused so individual test modules can depend on support without
+depending on another test module.
+"""
