+++ tests/test_input_dispatch_transfer.py
@@ -1,4 +1,4 @@
-from tests.test_state_reducer import _reduce_state
+from tests.support.state import reduce_state as _reduce_state
 from zivo.state import build_initial_app_state, dispatch_key_input
 from zivo.state.actions import (
     ActivateTabByIndex,
