+++ tests/test_state_reducer_palette_replace.py
@@ -1,6 +1,6 @@
 from dataclasses import replace
 
-from tests.state_test_helpers import reduce_state
+from tests.support.state import reduce_state
 from zivo.models import (
     TextReplacePreviewEntry,
     TextReplacePreviewResult,
