+++ tests/test_state_reducer_palette_search.py
@@ -4,7 +4,7 @@ from dataclasses import replace
 
 import pytest
 
-from tests.state_test_helpers import reduce_state
+from tests.support.state import reduce_state
 from zivo.models import (
     ExternalLaunchRequest,
 )
